+++ recfm-java-addon/src/main/java/io/github/epi155/recfm/cobol/CobolOccurs.java
@@ -1,0 +1,18 @@
+package io.github.epi155.recfm.cobol;
+
+public class CobolOccurs extends CobolEntry {
+    private final int times;
+
+    public CobolOccurs(int level, String name, int times) {
+        super(level, name);
+        this.times = times;
+    }
+
+    @Override
+    public String code() {
+        return spaceLevel() +
+                level() + " " +
+                name() +
+                " OCCURS " + times + " TIMES.";
+    }
+}