+++ recfm-java-addon/src/main/java/io/github/epi155/recfm/cobol/CobolPicX.java
@@ -1,0 +1,18 @@
+package io.github.epi155.recfm.cobol;
+
+public class CobolPicX extends CobolValue {
+    public CobolPicX(int level, String name, int offset, int length) {
+        super(level, name, offset, length);
+    }
+
+    @Override
+    public String code() {
+        return spaceLevel() +
+                level() + " " +
+                name() +
+                spacePic() +
+                " PIC X(" + length() +")."
+//                + tail()
+                ;
+    }
+}