+++ recfm-java-addon/src/main/java/io/github/epi155/recfm/java/fields/GroupTrait.java
@@ -1,6 +1,5 @@
 package io.github.epi155.recfm.java.fields;
 
-import io.github.epi155.recfm.api.FieldDefault;
 import io.github.epi155.recfm.java.factory.CodeWriter;
 import io.github.epi155.recfm.java.factory.DelegateWriter;
 import io.github.epi155.recfm.java.rule.InitializeField;
@@ -8,15 +7,13 @@ import io.github.epi155.recfm.type.FieldGroupTrait;
 import org.jetbrains.annotations.NotNull;
 
 public class GroupTrait extends DelegateWriter  implements InitializeField<FieldGroupTrait> {
-    private final FieldDefault defaults;
 
-    public GroupTrait(CodeWriter pw, FieldDefault defaults) {
+    public GroupTrait(CodeWriter pw) {
         super(pw);
-        this.defaults = defaults;
     }
 
     @Override
-    public void initialize(@NotNull FieldGroupTrait fld, int bias) {
+    public void initialize(@NotNull FieldGroupTrait fld) {
         if (! fld.isEmbedded())
             printf("    %s.initialize();%n", fld.getName());
     }