+++ recfm-java-addon/src/main/java/io/github/epi155/recfm/java/rule/InitializeField.java
@@ -3,7 +3,7 @@ package io.github.epi155.recfm.java.rule;
 import org.jetbrains.annotations.NotNull;
 
 public interface InitializeField<T> {
-    void initialize(@NotNull T fld, int bias);
+    void initialize(@NotNull T fld);
 
     void initializeItem(@NotNull T fld);
 }